move pending execution to transaction folder
